+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **`[fmtc]`** Fixed Off-by-One bug in 256-Color range validation
 - **`[timeutil]`** Fixed bug with formatting timezone offset
+- **`[timeutil]`** Fixed ISO week calculation
 - **`[fmtc]`** Code refactoring
 - **`[log]`** Code refactoring
 - **`[req]`** Code refactoring